extraído login do servidor para um módulo a parte
